Printf long hex

+++ Userland/Libc/include/stdlib.h
@@ -7,7 +7,7 @@
 
 extern int int80(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx);
 void * malloc(unsigned int size);
-int itoa(int value, char * buffer, int base);
+int itoa(uint64_t value, char * buffer, int base);
 int isAlpha(char c);
 int isDigit(char c);
 int rand();
