+++ Userland/Libc/stdlib.c
@@ -10,16 +10,16 @@ void * malloc(unsigned int size) {
 
 void free(void *ptr) {
 
-} 
+}
 
 //Adapted from naiveterminal.c
-int itoa(int value, char * buffer, int base) {
+int itoa(uint64_t value, char * buffer, int base) {
 	char *p = buffer;
 	char *p1, *p2;
 	int digits = 0;
 
 	if(value < 0 && base == 10) {
-		value = -value; 
+		value = -value;
 		*p++ = '-';
 		digits++;
 	}
@@ -56,7 +56,7 @@ int isAlpha(char c) {
 }
 
 int isDigit(char c) {
-	return (c >= '0' && c <= '9'); 
+	return (c >= '0' && c <= '9');
 }
 
 int rand(){
